% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azores_extent.R
\name{azores_extent}
\alias{azores_extent}
\title{Azores archipelago region}
\usage{
azores_extent()
}
\value{
A numeric vector of four elements:
\itemize{
\item \code{xmin}: most western longitude coordinate
\item \code{xmax}: most eastern longitude coordinate
\item \code{ymin}: most southern latitude coordinate
\item \code{ymax}: most northern latitude coordinate
}
}
\description{
\code{\link[=azores_extent]{azores_extent()}} returns the coordinates of the two vertices of the default
bounding box used to defined the Azores region in paikea studies.
}
