% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{azores_extent_utm}
\alias{azores_extent_utm}
\title{Azores archipelago region (N26 UTM)}
\usage{
azores_extent_utm()
}
\value{
A numeric vector of four elements:
\itemize{
\item \code{xmin}: most western UTM coordinate
\item \code{xmax}: most eastern UTM coordinate
\item \code{ymin}: most southern UTM coordinate
\item \code{ymax}: most northern UTM coordinate
}
}
\description{
\code{\link[=azores_extent_utm]{azores_extent_utm()}} returns the coordinates of the two vertices of the default
bounding box used to defined the Azores region in paikea studies.
}
