% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azores_grid.R
\name{azores_grid}
\alias{azores_grid}
\title{Creates an Azores grid}
\usage{
azores_grid(resolution = 500, type = c("rast", "stars"))
}
\arguments{
\item{resolution}{Resolution, defaults to 500 meters.}

\item{type}{The type of raster object returned: either a SpatRaster
(\code{"rast"}) or a stars (\code{"stars"}) object.}
}
\value{
A SpatRaster object.
}
\description{
\code{\link[=azores_grid]{azores_grid()}} creates a grid centered around the Azores region.
}
\details{
This region is defined as the bounding box of the Exclusive Economic Zone
(EEZ) for the Azores with an extra margin of approximately 50 km in each
direction.

This extent is for a custom Lambert Azimuthal Equal-Area (LAEA) projection,
centered on the Azores. See
\link[CAOP.RAA.2024:laea_azores_proj]{laea_azores_proj()} for more details about
this projection.
}
