% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{convert_extent_to_decimal}
\alias{convert_extent_to_decimal}
\title{Convert an extent in N26 UTM to WGS84}
\usage{
convert_extent_to_decimal(extent = azores_extent_utm())
}
\arguments{
\item{extent}{A named numeric vector of four N26 UTM values defining an
extent.}
}
\value{
A numeric vector.
}
\description{
\code{\link[=convert_extent_to_decimal]{convert_extent_to_decimal()}} takes a named numeric vector:
\itemize{
\item \code{xmin}: most western UTM coordinate
\item \code{xmax}: most eastern UTM coordinate
\item \code{ymin}: most southern UTM coordinate
\item \code{ymax}: most northern UTM coordinate
}

And generates a new one whose units have been converted from N26 UTM to WGS84
(decimal degrees).
}
