% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{convert_extent_to_utm}
\alias{convert_extent_to_utm}
\title{Convert an extent in WGS84 to N26 UTM}
\usage{
convert_extent_to_utm(extent = azores_extent_decimal())
}
\arguments{
\item{extent}{A named numeric vector of four WGS84 values defining an extent.}
}
\value{
A numeric vector.
}
\description{
\code{\link[=convert_extent_to_utm]{convert_extent_to_utm()}} takes a named numeric vector:
\itemize{
\item \code{xmin}: most western longitude coordinate, in decimal degrees.
\item \code{xmax}: most eastern longitude coordinate, in decimal degrees.
\item \code{ymin}: most southern latitude coordinate, in decimal degrees.
\item \code{ymax}: most northern latitude coordinate, in decimal degrees.
}

And generates a new one whose units have been converted from WGS84 to N26 UTM
(meters).
}
