% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathymetry.R
\name{bathymetry}
\alias{bathymetry}
\title{Azores bathymetry, depth and elevation}
\usage{
bathymetry(
  var = c("bathymetry", "depth", "elevation", "depth_slope"),
  resolution = 500
)
}
\arguments{
\item{var}{Either \code{"bathymetry"}, \code{"depth"}, \code{"elevation"} or
\code{"depth_slope"}.}

\item{resolution}{A grid cell size, in meters. Can be a single value,
specifying the size for both the x-dimension (longitude) and y-dimension
(latitude), or a vector of two values for an explicit specification. By
default the original resolution corresponds to 15 arcsecs.}
}
\value{
A SpatRaster object.
}
\description{
\code{\link[=bathymetry]{bathymetry()}} creates a SpatRaster object for the Azores region with either
bathymetry, depth or elevation values (in meters). Use the \code{var} parameter to
choose the variable of interest. The area returned in for the Azores,
covering the EEZ plus an extra margin of 50 km on each side, see
\code{\link[=azores_extent]{azores_extent()}} for more details.
}
