% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathymetry.R
\name{bathymetry}
\alias{bathymetry}
\title{Azores bathymetry, depth and elevation}
\usage{
bathymetry(
  var = c("bathymetry", "depth", "elevation", "depth_slope"),
  extent = azores_extent(),
  resolution = 1000
)
}
\arguments{
\item{var}{Either \code{"bathymetry"}, \code{"depth"}, \code{"elevation"} or
\code{"depth_slope"}.}

\item{extent}{The extent of the region to be selected. This should be a
numeric vector of four values:
\itemize{
\item \code{xmin}: most western longitude coordinate, in decimal degrees.
\item \code{xmax}: most eastern longitude coordinate, in decimal degrees.
\item \code{ymin}: most southern latitude coordinate, in decimal degrees.
\item \code{ymax}: most northern latitude coordinate, in decimal degrees.
}

Defaults to large area centered at the Azores archipelago:
-34, -21, 34, 42.

Note that indicating an extent beyond the default limits will not yield
more data points.}

\item{resolution}{A grid cell size, in meters. Can be a single value,
specifying the size for both the x-dimension (longitude) and y-dimension
(latitude), or a vector of two values for an explicit specification.}
}
\value{
A SpatRaster object cropped to the region selected with \code{extent},
whose grid cells have a resolution as indicated in the \code{resolution} parameter
}
\description{
\code{\link[=bathymetry]{bathymetry()}} creates a SpatRaster object for the Azores region with either
bathymetry, depth or elevation values (in meters). Use the \code{var} parameter to
choose the variable of interest.
}
